sql_ops_agent/eval/harness.py
import yaml
import asyncio
from pathlib import Path
from dataclasses import dataclass
from typing import List, Any
# Matches the Orchestrator interface we will build
# from sql_ops_agent.orchestrator import AgentOrchestrator

@dataclass
class CaseResult:
    case_id: str
    passed: bool
    details: str

class EvalHarness:
    def __init__(self, cases_path: Path):
        self.cases = yaml.safe_load(cases_path.read_text())
        self.results: List[CaseResult] = []

    async def run(self):
        print(f"Running {len(self.cases)} cases...")
        for case in self.cases:
            # Placeholder for actual agent call
            # result = await agent.run(case['prompt'])
            # check correctness
            pass
        
        # Determine pass/fail based on expected sql/behavior
        # This is a skeleton as requested.
        print("Eval run complete.")

if __name__ == "__main__":
    h = EvalHarness(Path("bench/cases.yaml"))
    asyncio.run(h.run())
