pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sql-ops-agent"
version = "0.1.0"
description = "Local-first SQL Ops Agent with guardrails and RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy[asyncio]",
    "aiosqlite",  # For DuckDB async compatibility if needed, though duckdb has its own
    "duckdb",
    "duckdb-engine",
    "asyncpg",
    "sqlglot",
    "httpx",
    "structlog",
    "prometheus-client",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "rank_bm25", # For simple RAG
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "types-pyyaml",
]

[tool.hatch.build.targets.wheel]
packages = ["sql_ops_agent"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
